refactor settings & manage navigation stack
