+++ clients/flutter_client/lib/src/widgets/settings/signed_in_settings_view.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_client/src/controllers/settings_controller.dart';
+import 'package:flutter_client/src/widgets/settings/about_view.dart';
+import 'package:flutter_client/src/widgets/settings/logout_button.dart';
+import 'package:flutter_client/src/widgets/settings/require_auth_setting_view.dart';
+import 'package:flutter_client/src/widgets/settings/theme_setting_view.dart';
+
+class SignedInSettingsView extends StatelessWidget {
+  const SignedInSettingsView({super.key, required this.settingsController});
+
+  final SettingsController settingsController;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Settings'),
+      ),
+      body: Column(children: [
+        ThemeSettingView(settingsController),
+        RequireAuthSettingView(settingsController),
+        AboutView(
+          appName: settingsController.appName,
+          appVersion: settingsController.appVersion,
+        ),
+      ]),
+      bottomNavigationBar: LogoutButton(controller: settingsController),
+    );
+  }
+}
