+++ clients/flutter_client/lib/src/widgets/settings/theme_setting_view.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_client/src/controllers/settings_controller.dart';
+import 'package:flutter_client/src/widgets/menus/theme_menu.dart';
+
+class ThemeSettingView extends StatelessWidget {
+  const ThemeSettingView(
+    this.controller, {
+    super.key,
+  });
+
+  final SettingsController controller;
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: const Text('Theme'),
+      trailing: ThemeMenu(controller: controller),
+    );
+  }
+}
